requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cleanup==9.0.0
django-formtools==2.5.1
pillow==11.0.0
sqlparse==0.5.1
tzdata==2024.2
